Refactor GetId method to change ID string format
The GetId method in the GitHubCopilotAssignedSeats class has been
updated to change the format of the returned ID strings.
The new format places the day at the beginning, followed by a hyphen, the
type of entity (ORG, ENT, or XXX), and then the organization or
enterprise name if applicable.
This change affects all three possible return values of the method.